+++ src/main/java/com/community/community/post/PostService.java
@@ -11,7 +11,6 @@ import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
 import java.util.List;
-import java.util.Optional;
 
 import static com.community.community.post.PostDto.*;
 
@@ -24,16 +23,16 @@ public class PostService {
     private final UserRepository userRepository;
     private final CommentRepository commentRepository;
 
-    public Long writePost(PostCreateRequest request) {
+    public Long writePost(PostCreateRequest request, Long userId) {
         // 1. 글을 작성하는 사용자를 DB에서 조회
-        UserEntity user = userRepository.findById(request.userId())
+        UserEntity findUser = userRepository.findById(userId)
                 .orElseThrow(() -> new IllegalArgumentException("존재하지 앟는 회원입니다."));
 
-        // 2. DTO의 User를 조합하여 PostEntity 만듬
+        // 2. DTO의 User를 PostEntity 조합
         PostEntity post = PostEntity.builder()
                 .title(request.title())
                 .content(request.content())
-                .userEntity(user) // 게시글의 글쓴이와 연결
+                .userEntity(findUser) // 유저 객체 맵핑
                 .build();
 
         // 3. DB에 저장
@@ -55,23 +54,31 @@ public class PostService {
         return PostDetailResponse.from(post, comments);
     }
 
-    public Long updatePost(Long id, PostUpdateRequest request) {
+    public void updatePost(Long id, PostUpdateRequest request, Long userId) {
         // 1. 수정할 게시글을 DB에서 조회
         PostEntity post = postRepository.findById(id)
                 .orElseThrow(() -> new IllegalArgumentException("존재하지 않는 게시글입니다."));
 
-        // 2. 찾아온 객체의 값을 수정
-        post.update(request.title(), request.content());
+        // 2. 작성자와 수정 요청자 일치 검증
+        if (!post.getUserEntity().getId().equals(userId)) {
+            throw new IllegalArgumentException("작성자만 수정할 수 있습니다.");
+        }
 
-        return post.getId();
+        // 3. 통과시 수정
+        post.update(request.title(), request.content());
     }
 
-    public void deletePost(Long id) {
+    public void deletePost(Long id,Long userId) {
         // 1. 삭제할 게시글을 DB에서 조회
         PostEntity post = postRepository.findById(id)
                 .orElseThrow(() -> new IllegalArgumentException("존재하지 않는 게시글입니다."));
 
-        // 2. 찾은 게시글을 DB에서 삭제.
+        // 2. 작성자와 삭제 요청자 일치 검증
+        if (!post.getUserEntity().getId().equals(userId)) {
+            throw new IllegalArgumentException("작성자만 삭제할 수 있습니다.");
+        }
+
+        // 3. 찾은 게시글을 DB에서 삭제.
         postRepository.delete(post);
     }
